requirements.txt
streamlit==1.40.0
pandas>=2.1.0
openpyxl>=3.1.2
dashscope>=1.13.0
python-dotenv>=1.0.0
